# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a todo API.",
        "title": "TodoAPI",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:1006",
    "basePath": "/api/v1",
    "paths": {
        "/": {
            "post": {
                "description": "Use to filter todo by title, status, due date, priority",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Filter Todo",
                "parameters": [
                    {
                        "description": "Todo title, status, dateFrom, dateStart, priority",
                        "name": "filterData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/checkToken": {
            "get": {
                "description": "Use to check jwt token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Check Valid Token",
                "responses": {}
            }
        },
        "/create": {
            "post": {
                "description": "Use to create todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Create Todo",
                "parameters": [
                    {
                        "description": "Todo title, description, dueDate, priority",
                        "name": "createData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/delete": {
            "delete": {
                "description": "Use to delete todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Delete Todo",
                "parameters": [
                    {
                        "description": "Todo todoId",
                        "name": "deleteData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/login": {
            "post": {
                "description": "Login todo website by name and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "User name and password",
                        "name": "name\u0026password",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/register": {
            "post": {
                "description": "Register todo website by name, password and rePassword",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "User name, password, rePassword",
                        "name": "name\u0026password\u0026rePassword",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/update": {
            "patch": {
                "description": "Use to update todo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todos"
                ],
                "summary": "Update Todo",
                "parameters": [
                    {
                        "description": "Todo todoId, updateField, updateValue",
                        "name": "updateData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "securityDefinitions": {
        "JWT": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}